pyproject.toml
[tool.poetry]
name = "image-processing-function-app"
version = "0.1.0"
description = "The image-processing-function-app is a Python package designed to provide a collection of functions for image processing tasks"
authors = ["rterluun <rterluun@ilionx.com>"]
readme = "README.md"
packages = [
    { include = "image_processing_function_app", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
azure-functions = "^1.19.0"
exif = "^1.6.0"
azure-storage-blob = "^12.19.1"
azure-identity = "^1.16.0"
azure-data-tables = "^12.5.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
pytest = "^8.1.1"
coverage = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = [
  "v1", "src",
]

[tool.isort]
profile = "black"
